pyproject.toml
[project]
name = "ai-chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.2",
    "groq>=0.37.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "pydantic>=2.12.5",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
